+++ ndk/src/media_error.rs
@@ -1,7 +1,8 @@
 //! Bindings for NDK media status codes.
 //!
 //! Also used outside of `libmediandk.so` in `libamidi.so` for example.
-#![cfg(feature = "media")]
+
+#![cfg(any(feature = "media", feature = "midi"))]
 // The cfg(feature) bounds for some pub(crate) fn uses are non-trivial and will become even more
 // complex going forward.  Allow them to be unused when compiling with certain feature combinations.
 #![allow(dead_code)]